+++ source/src/Peaks.cpp
@@ -9,6 +9,7 @@
 // ============================================================================
 #include "gsl/gsl_sf_exp.h"
 #include "gsl/gsl_sf_gamma.h"
+#include "gsl/gsl_sf_gamma.h"
 #include "gsl/gsl_randist.h"
 #include "gsl/gsl_cdf.h"
 // ============================================================================
@@ -19,6 +20,7 @@
 // ============================================================================
 //  Local
 // ============================================================================
+#include "Exception.h"
 #include "local_math.h"
 #include "local_gsl.h"
 #include "local_hash.h"
@@ -3346,9 +3348,6 @@ std::size_t Ostap::Math::RaisingCosine::tag () const
 // ============================================================================
 
 
-
-
-
 // ============================================================================
 /*  constructor from all parameters 
  *  @param mu  peak location
@@ -3605,7 +3604,178 @@ double Ostap::Math::QGaussian::integral ( const double low  ,
 std::size_t Ostap::Math::QGaussian::tag () const 
 { return std::hash_combine ( m_mean , m_q , m_scale ) ; }
 // ============================================================================
- 
+
+// ============================================================================
+/* constructor from mu, sigma, zeta and kappa 
+ *  @param mu    related to location 
+ *  @param beta  related to asymmetry
+ *  @param sigma related to width 
+ *  @param zeta  related to what ?
+ */
+// ============================================================================
+Ostap::Math::Hyperbolic::Hyperbolic
+( const double mu     ,   // related to location 
+  const double sigma  ,   // related to withs  
+  const double zeta   ,   // shape parameter
+  const double kappa  )   // related to asymmetry 
+  : m_mu    ( mu    ) 
+  , m_sigma ( -1    )
+  , m_zeta  ( -1    )
+  , m_kappa ( kappa )
+  , m_A     (  0    )  
+  , m_K1    (  0    )
+{
+  setSigma ( sigma ) ;
+  setZeta  ( zeta  ) ;
+}
+// ============================================================================
+bool Ostap::Math::Hyperbolic::setMu    ( const double value ) 
+{
+  if ( s_equal ( value , m_mu  ) ) { return false ; }
+  m_mu = value ;
+  return true ;
+}
+// ============================================================================
+bool Ostap::Math::Hyperbolic::setSigma ( const double value ) 
+{
+  const double avalue = std::abs ( value ) ;
+  if ( s_equal ( avalue , m_sigma ) ) { return false ; }
+  m_sigma = avalue ;
+  return true ;
+}
+// ============================================================================
+bool Ostap::Math::Hyperbolic::setZeta ( const double value ) 
+{
+  const double avalue = std::abs ( value ) ;
+  if ( s_equal ( avalue , m_zeta ) ) { return false ; }
+  m_zeta = avalue ;
+  //
+  if ( m_zeta > 1.e-7 )
+  {
+    m_K1 = m_zeta                  * Ostap::Math::bessel_Kn_scaled ( 1 , m_zeta )          ;
+    m_A  = std::pow ( m_zeta , 2 ) * Ostap::Math::bessel_Kn_scaled ( 2 , m_zeta ) / m_K1   ;
+  }
+  else 
+  {
+    m_K1 = 1 ;
+    m_A  = 2 ;
+  }
+  //
+  return true ;
+}
+// ============================================================================
+bool Ostap::Math::Hyperbolic::setKappa ( const double value ) 
+{
+  if ( s_equal ( value , m_kappa ) ) { return false ; }
+  m_kappa = value ;
+  return true ;
+}
+// ============================================================================
+// calculate the mean of the distribution  
+// ============================================================================
+double Ostap::Math::Hyperbolic::mean () const 
+{ return m_mu + m_kappa * m_sigma ; }
+// ============================================================================
+// get the actual mode of the distribution
+// ============================================================================
+double Ostap::Math::Hyperbolic::mode () const 
+{ return m_mu + m_kappa * m_sigma * m_zeta / m_A ; }
+// ============================================================================
+// get the variance/dispersion 
+// ============================================================================
+double Ostap::Math::Hyperbolic::variance () const 
+{
+  double var = m_sigma * m_sigma ;
+  //
+  if ( s_zero ( m_kappa ) ) { return var ; }  // RETURN 
+  //
+  const double beta2  = std::pow ( beta () , 2 ) ;
+  if ( s_zero ( beta2   ) ) { return var ; }  // RETURN 
+  //
+  const double gamma4 = std::pow ( gamma2() , 2 ) ;
+  //
+  const double t1 = ( 1.e-7 > m_zeta ) ? 8.0 : 
+    std::pow ( m_zeta , 3 ) * Ostap::Math::bessel_Kn_scaled ( 3 , m_zeta ) / m_K1 ;
+  //
+  const double t2 = m_A * m_A ;
+  //
+  var += beta2 / gamma4 * ( t1 - t2 ) ;
+  //
+  return var;
+}
+// ============================================================================
+// evaluate  pdf  for the Hyperbolic distribution
+// ============================================================================
+double Ostap::Math::Hyperbolic::pdf ( const double x ) const 
+{
+  //
+  const double dx =  ( x - m_mu ) / m_sigma ;
+  //
+  const double cc = m_kappa * m_kappa + m_A ;
+  //
+  const double q  = 
+    m_zeta - std::sqrt ( cc * m_zeta * m_zeta / m_A + cc * dx * dx ) + m_kappa * dx  ;
+  //
+  return std::exp ( q ) * m_A / ( 2 * m_sigma * std::sqrt ( cc ) * m_K1 ) ;
+}
+// ============================================================================
+// get the integral between low and high limits
+// =========================================================================
+double Ostap::Math::Hyperbolic::integral
+( const double low  ,
+  const double high ) const
+{
+  //
+  if      ( s_equal ( low , high ) ) { return                 0.0        ; } // RETURN
+  else if (           low > high   ) { return - integral ( high , low  ) ; } // RETURN
+  //
+  const double m1    = mode () ;
+  const double m2    = mean () ;
+  const double mmin  = std::min ( m1 , m2 ) ;
+  const double mmax  = std::max ( m1 , m2 ) ;
+  const double mlow  = mmin - 5 * m_sigma ;
+  const double mhigh = mmax + 5 * m_sigma ;
+  //
+  const double mc [] = { mmin - 3.0 * m_sigma , 
+                         mmax + 3.0 * m_sigma , 
+                         mlow , mhigh         } ;
+  //
+  for ( const double c : mc ) 
+  { if ( low < c  && c < high ) { return integral ( low , c ) + integral ( c , high ) ; } }
+  //
+  // in tails 
+  const bool in_tail = ( high <= mlow ) || ( low >= mhigh ) ;
+  //
+  // use GSL to evaluate the integral
+  //
+  static const Ostap::Math::GSL::Integrator1D<Hyperbolic> s_integrator {} ;
+  static char s_message[] = "Integral(Hyperbolic)" ;
+  //
+  const auto F = s_integrator.make_function ( this ) ;
+  int    ierror   =  0 ;
+  double result   =  1 ;
+  double error    = -1 ;
+  std::tie ( ierror , result , error ) = s_integrator.gaq_integrate
+    ( tag () , 
+      &F     ,  
+      low    , high  ,               // low & high edges
+      workspace ( m_workspace ) ,    // workspace
+      in_tail ? s_PRECISION_TAIL : s_PRECISION , // absolute precision
+      in_tail ? s_PRECISION_TAIL : s_PRECISION , // relative precision
+      m_workspace.size () ,          // size of workspace
+      s_message           , 
+      __FILE__ , __LINE__ ) ;
+  //
+  return result ;
+  //
+}
+// ============================================================================
+// get the tag 
+// ============================================================================
+std::size_t Ostap::Math::Hyperbolic::tag () const 
+{ return std::hash_combine ( m_mu , m_sigma , m_zeta , m_kappa  ) ; }
+// ============================================================================
+
 
 // ============================================================================
 //                                                                      The END